requirements.txt
slack-bolt==1.18.0
python-dotenv==1.0.0
uvicorn==0.29.0
redis==5.0.1
openai~=1.0
pymupdf~=1.24.0
httpx==0.27.0
pytest-asyncio~=0.21.0
pytest-cov~=4.1.0
aiohttp 